Add user idling feature
If no input or mouse movement is detected within a set amount of time while working (8 seconds by default), the work timer will pause and indicate that you are currently idling and the amount of time that you've been idling for.

The time spent idling will not be counted towards how long you've been working.